requirements.txt
streamlit
pandas
cryptography
python-dotenv
langchain
langchain-openai
